Folders cleaned up further. Layout created, remember to explain in the morning. Not insistent on anything, please edit any updates ive made as you wish.
