+++ next.config.js
@@ -23,7 +23,7 @@ const nextConfig = {
   //   return [
   //     {
   //       source: "/api/:path*",
-  //       destination: "http://localhost:3000/:path*",
+  //       destination: "http://localhost:3000/api/:path*",
   //     },
   //   ];
   // },
