+++ package.json
@@ -10,8 +10,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "seed": "npx prisma migrate dev",
-    "reseed": "yarn prisma migrate reset"
+    "migrate": "yarn prisma migrate dev",
+    "seed": "yarn prisma migrate reset"
   },
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.5",
