+++ src/pages/api/auth/[...nextauth].ts
@@ -24,9 +24,10 @@ export default NextAuth({
       }
       return token;
     },
-    // async session({ session, token }) {
+    // async session({ session, token, user }) {
     //   // Send properties to the client, like an access_token from a provider.
-    //   session.accessToken = token.accessToken;
+    //   // session.accessToken = token.accessToken;
+    //   session.user.isAdmin = user.isAdmin;
     //   return session;
     // },
   },
