+++ src/pages/test.tsx
@@ -1,4 +1,4 @@
-import AllPosts from "@/components/allPosts";
+import AllPosts from "@/components/PostComponents/allPosts";
 
 export default function Test() {
   return <div className="bg-white h-screen">test </div>;
